Assignment 1
Assignment 1 consists of a web server for working with an xlsx file (a kind of API) containing information about books, authors and genres (all matches with real people and books are random).
The purpose of the assignment was to implement four types of requests: GET, POST, PUT and DELETE. As well as the corresponding CRUD implementation
The system also has validation for entering new data, pagination for convenient data display, writing and reading an xlsx file. 
The MVC pattern was implemented for convenient data management.